pyproject.toml
[project]
name = "DeskPilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.5",
    "google-generativeai",
    "pyyaml",
    "pillow",
    "pyautogui",
    "dspy",
    "litellm",
]
